+++ src/ModularPipelines/Engine/Executors/IExecutionOrchestrator.cs
@@ -1,0 +1,8 @@
+﻿using ModularPipelines.Modules;
+
+namespace ModularPipelines.Engine.Executors;
+
+internal interface IExecutionOrchestrator
+{
+    Task<IReadOnlyList<ModuleBase>> ExecuteAsync();
+}