+++ src/ModularPipelines/Engine/Executors/IModuleDisposeExecutor.cs
@@ -1,0 +1,8 @@
+﻿using ModularPipelines.Modules;
+
+namespace ModularPipelines.Engine.Executors;
+
+internal interface IModuleDisposeExecutor
+{
+    Task ExecuteAndDispose(IEnumerable<ModuleBase> modules, Func<Task> executeDelegate);
+}