+++ src/ModularPipelines/Engine/Executors/IPrintModuleOutputExecutor.cs
@@ -1,0 +1,8 @@
+﻿using ModularPipelines.Modules;
+
+namespace ModularPipelines.Engine.Executors;
+
+internal interface IPrintModuleOutputExecutor
+{
+    Task ExecuteAndPrintModuleOutput(Func<Task> executeDelegate);
+}