+++ src/ModularPipelines/Engine/Executors/IPrintProgressExecutor.cs
@@ -1,0 +1,9 @@
+﻿using ModularPipelines.Models;
+using ModularPipelines.Modules;
+
+namespace ModularPipelines.Engine.Executors;
+
+internal interface IPrintProgressExecutor
+{
+    Task ExecuteWithProgress(OrganizedModules organizedModules, Func<Task> executeDelegate);
+}