+++ src/ModularPipelines/Engine/ILogoPrinter.cs
@@ -1,0 +1,6 @@
+﻿namespace ModularPipelines.Engine;
+
+internal interface ILogoPrinter
+{
+    void PrintLogo();
+}