+++ src/ModularPipelines/Engine/IPipelineExecutor.cs
@@ -1,8 +1,0 @@
-using ModularPipelines.Modules;
-
-namespace ModularPipelines.Engine;
-
-public interface IPipelineExecutor
-{
-    Task<IReadOnlyList<ModuleBase>> ExecuteAsync();
-}