+++ src/ModularPipelines/Helpers/IConsolePrinter.cs
@@ -1,0 +1,7 @@
+﻿using ModularPipelines.Engine;
+
+namespace ModularPipelines.Helpers;
+
+internal interface IConsolePrinter : IProgressPrinter, ILogoPrinter, IDependencyPrinter
+{
+}