+++ src/ModularPipelines/Helpers/IProgressPrinter.cs
@@ -2,7 +2,9 @@ using ModularPipelines.Models;
 
 namespace ModularPipelines.Helpers;
 
-internal interface IPipelineConsolePrinter
+internal interface IProgressPrinter
 {
     Task PrintProgress(OrganizedModules organizedModules, CancellationToken cancellationToken);
+
+    void PrintResults(PipelineSummary pipelineSummary);
 }