+++ src/ModularPipelines/Helpers/PathType.cs
@@ -1,0 +1,7 @@
+﻿namespace ModularPipelines.Helpers;
+
+internal enum PathType
+{
+    Directory,
+    File
+}